requirements.txt
# HTTP requests
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Date handling (used in crypto_tool.py)
python-dateutil>=2.8.2

# Type hints (used throughout the codebase)
typing-extensions>=4.9.0

# Optional but recommended for development
pytest>=8.0.0  # For testing
black>=24.1.0  # For code formatting
flake8>=7.0.0  # For linting
